Ajustes nos Controllers e Criação de Testes

+++ src/main/java/br/com/fiap/PetPass/repository/EstabelecimentoRepository.java
@@ -6,7 +6,7 @@ import br.com.fiap.PetPass.entity.Estabelecimento;
 
 public interface EstabelecimentoRepository extends JpaRepository<Estabelecimento, Long> {
 	
-	Estabelecimento findByDocumento(Integer documento);
+	Estabelecimento findByDocumento(Long documento);
 	
 	Estabelecimento findByNome(String nome);
